More Progress
Re-textured the ground and hills in the egg game. Finished level layout and animation for the marble game (at this point I may not even bother with trying to re-work the movement style and camera , we can discuss it later today or tomorrow). Added a baby dragon as the secondary player model for the infinite jumper; However the viking hat doesn't load in on launch and the game thinks it's still there even when it's "unequipped". We do, however, still need to set-up the main menu as well as add buttons to exit the games and the collection.